[BACKEND] Refactoring, divide HashGenerator to two other classes
